Add files via upload
get_next_line consists of creating a function that reads a file descriptor line by line and returns the next line each time it is called. The bonus part of this project requires altering the function so that it can handle multiple file descriptors simultaneously. This project aims to reinforce concepts such as memory management, file descriptors, and string manipulation.